+++ lib/ansible/modules/storage/netapp/netapp_e_hostgroup.py
@@ -31,30 +31,9 @@ short_description: Manage NetApp Storage Array Host Groups
 author: Kevin Hulquest (@hulquest)
 description:
 - Create, update or destroy host groups on a NetApp E-Series storage array.
+extends_documentation_fragment:
+    - netapp.eseries
 options:
-  api_username:
-      required: true
-      description:
-      - The username to authenticate with the SANtricity WebServices Proxy or embedded REST API.
-  api_password:
-      required: true
-      description:
-      - The password to authenticate with the SANtricity WebServices Proxy or embedded REST API.
-  api_url:
-      required: true
-      description:
-      - The url to the SANtricity WebServices Proxy or embedded REST API.
-      example:
-      - https://prod-1.wahoo.acme.com/devmgr/v2
-  validate_certs:
-      required: false
-      default: true
-      description:
-      - Should https certificates be validated?
-  ssid:
-    required: true
-    description:
-    - The ID of the array to manage (as configured on the web services proxy).
   state:
     required: true
     description:
@@ -137,44 +116,11 @@ HEADERS = {
 
 import json
 
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.pycompat24 import get_exception
-
-from ansible.module_utils.urls import open_url
 from ansible.module_utils.six.moves.urllib.error import HTTPError
 
-
-def request(url, data=None, headers=None, method='GET', use_proxy=True,
-            force=False, last_mod_time=None, timeout=10, validate_certs=True,
-            url_username=None, url_password=None, http_agent=None, force_basic_auth=True, ignore_errors=False):
-    try:
-        r = open_url(url=url, data=data, headers=headers, method=method, use_proxy=use_proxy,
-                     force=force, last_mod_time=last_mod_time, timeout=timeout, validate_certs=validate_certs,
-                     url_username=url_username, url_password=url_password, http_agent=http_agent,
-                     force_basic_auth=force_basic_auth)
-    except HTTPError:
-        err = get_exception()
-        r = err.fp
-
-    try:
-        raw_data = r.read()
-        if raw_data:
-            data = json.loads(raw_data)
-        else:
-            raw_data = None
-    except:
-        if ignore_errors:
-            pass
-        else:
-            raise Exception(raw_data)
-
-    resp_code = r.getcode()
-
-    if resp_code >= 400 and not ignore_errors:
-        raise Exception(resp_code, data)
-    else:
-        return resp_code, data
-
+from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils.netapp import request, eseries_host_argument_spec
+from ansible.module_utils.pycompat24 import get_exception
 
 def group_exists(module, id_type, ident, ssid, api_url, user, pwd):
     rc, data = get_hostgroups(module, ssid, api_url, user, pwd)
@@ -326,19 +272,16 @@ def get_hosts_in_group(module, ssid, group_name, api_url, user, pwd):
 
 
 def main():
+    argument_spec = eseries_host_argument_spec()
+    argument_spec = argument_spec.update(dict(
+        name=dict(required=False),
+        new_name=dict(required=False),
+        id=dict(required=False),
+        state=dict(required=True, choices=['present', 'absent']),
+        hosts=dict(required=False, type='list'),
+    ))
     module = AnsibleModule(
-        argument_spec=dict(
-            name=dict(required=False),
-            new_name=dict(required=False),
-            ssid=dict(required=True),
-            id=dict(required=False),
-            state=dict(required=True, choices=['present', 'absent']),
-            hosts=dict(required=False, type='list'),
-            api_url=dict(required=True),
-            api_username=dict(required=True),
-            validate_certs=dict(required=False, default=True),
-            api_password=dict(required=True, no_log=True)
-        ),
+        argument_spec=argument_spec,
         supports_check_mode=False,
         mutually_exclusive=[['name', 'id']],
         required_one_of=[['name', 'id']]
